backend/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
websockets>=12.0
mediapipe>=0.10.21
opencv-contrib-python==4.8.1.78
numpy>=1.24.0,<2.0
Pillow>=10.0.0
python-multipart>=0.0.6